+++ sezzle-sdk/src/main/kotlin/com/sezzle/sdk/checkout/SezzleCheckoutWebViewActivity.kt
@@ -197,7 +197,6 @@ class SezzleCheckoutWebViewActivity : Activity() {
             insets
         }
 
-        SezzleSessionScrubber.clear()
         webView.loadUrl(urlWithParam)
     }